Move source code to src directory
It is frequently recommended but I never felt a strong reason to use a separate src directory.
After working with a couple of projects, the 'best' reason for me is to allow my muscle-memory to 'remember' where the source directory is within the root directory.

Due to update to pdm-bump, I do not need to convert the bump type anymore.

Updated the pre-commit files.
Removed the default version constraints from dev dependencies